Fix 64-bit warnings (for 64-bit Windows build)
BUG=28176
TEST=none
Review URL: http://codereview.chromium.org/416003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@32620 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 7dc8d79bc15a895f02aa720dc6aa5af190adc678

+++ ipc/ipc_sync_channel.cc
@@ -422,7 +422,7 @@ void SyncChannel::WaitForReply(WaitableEvent* pump_messages_event) {
     };
 
     unsigned count = pump_messages_event ? 3: 2;
-    unsigned result = WaitableEvent::WaitMany(objects, count);
+    size_t result = WaitableEvent::WaitMany(objects, count);
     if (result == 0 /* dispatch event */) {
       // We're waiting for a reply, but we received a blocking synchronous
       // call.  We must process it or otherwise a deadlock might occur.
